♻️ 1차 리팩토링
