+++ .prettierrc
@@ -7,6 +7,7 @@
     "^@apis/(.*)$",
     "^@assets/(.*)$",
     "^@components/(.*)$",
+    "^@layouts/(.*)$",
     "^@constants/(.*)$",
     "^@hooks/(.*)$",
     "^@pages/(.*)$",
