+++ package.json
@@ -15,7 +15,7 @@
     "format": "eslint src --ext .ts,.tsx --fix && prettier --write \"src/**/*.{ts,tsx}\""
   },
   "dependencies": {
-    "@tanstack/react-virtual": "^3.0.0-alpha.0",
+    "@tanstack/react-virtual": "beta",
     "axios": "^1.4.0",
     "electron-is-dev": "^2.0.0",
     "firebase": "^9.23.0",
