+++ src/layouts/PageLayout.ts
@@ -3,7 +3,6 @@ import styled from "styled-components";
 export default styled.div`
   width: 754px;
   height: calc(100vh - 38px);
-  overflow-y: auto;
 
   margin-left: auto;
   margin-right: auto;
