+++ src/pages/user/Playlists.tsx
@@ -3,9 +3,10 @@ import { useCallback, useEffect, useReducer, useState } from "react";
 import { useRecoilState } from "recoil";
 import styled from "styled-components";
 
-import DefaultScroll from "@components/globals/Scroll/DefaultScroll";
 import Menu from "@components/user/playlist/Menu";
 import PlaylistItem from "@components/user/playlist/PlaylistItem";
+
+import PageItemContainer from "@layouts/PageItemContainer";
 
 import colors from "@constants/colors";
 import { myList } from "@constants/dummys";
@@ -129,59 +130,58 @@ const Playlists = ({}: PlaylistsProps) => {
   return (
     <Container>
       <Menu />
-      <DefaultScroll>
-        <ScrollWrapper>
-          <PlayLists
-            onMouseMove={mouseDown && isEditMode ? movePlayList : undefined}
-            onMouseUp={() => {
-              setMouseDown(false);
-              if (!mouseDown) return;
 
-              dispatchMyList({
-                type: ShuffleActionType.relocate,
-                target: dragAndDropTarget.drag.index,
-                to: dragAndDropTarget.drop,
-              });
-              // API에 내 리스트 수정 요청 보내기
-            }}
-            onMouseLeave={() => {
-              setMouseDown(false);
+      <PageItemContainer hasButton>
+        <PlayLists
+          onMouseMove={mouseDown && isEditMode ? movePlayList : undefined}
+          onMouseUp={() => {
+            setMouseDown(false);
+            if (!mouseDown) return;
+
+            dispatchMyList({
+              type: ShuffleActionType.relocate,
+              target: dragAndDropTarget.drag.index,
+              to: dragAndDropTarget.drop,
+            });
+            // API에 내 리스트 수정 요청 보내기
+          }}
+          onMouseLeave={() => {
+            setMouseDown(false);
+          }}
+        >
+          {!isEditMode
+            ? myList.map((item, index) => (
+                <PlaylistItem
+                  key={index}
+                  item={{
+                    ...item,
+                    index: index,
+                  }}
+                />
+              ))
+            : shuffledList.map((item, index) => (
+                <PlaylistItem
+                  key={index}
+                  item={{
+                    ...item,
+                    index: index,
+                  }}
+                  hide={index === dragAndDropTarget.drag.index && mouseDown}
+                  mouseDown={mouseDown}
+                  onSelect={initializeDragTarget}
+                />
+              ))}
+          <DragedPlaylist
+            style={{
+              top: `${dragPosition.y}px`,
+              left: `${dragPosition.x}px`,
+              display: mouseDown ? "block" : "none",
             }}
           >
-            {!isEditMode
-              ? myList.map((item, index) => (
-                  <PlaylistItem
-                    key={index}
-                    item={{
-                      ...item,
-                      index: index,
-                    }}
-                  />
-                ))
-              : shuffledList.map((item, index) => (
-                  <PlaylistItem
-                    key={index}
-                    item={{
-                      ...item,
-                      index: index,
-                    }}
-                    hide={index === dragAndDropTarget.drag.index && mouseDown}
-                    mouseDown={mouseDown}
-                    onSelect={initializeDragTarget}
-                  />
-                ))}
-            <DragedPlaylist
-              style={{
-                top: `${dragPosition.y}px`,
-                left: `${dragPosition.x}px`,
-                display: mouseDown ? "block" : "none",
-              }}
-            >
-              <PlaylistItem item={dragAndDropTarget.drag} />
-            </DragedPlaylist>
-          </PlayLists>
-        </ScrollWrapper>
-      </DefaultScroll>
+            <PlaylistItem item={dragAndDropTarget.drag} />
+          </DragedPlaylist>
+        </PlayLists>
+      </PageItemContainer>
     </Container>
   );
 };
@@ -199,13 +199,9 @@ const PlayLists = styled.div`
   align-content: flex-start;
   gap: 16px;
 
-  width: 100%;
+  /* width: 100%; */
 
   padding-right: 2px;
-`;
-
-const ScrollWrapper = styled.div`
-  height: calc(100vh - 206px);
 `;
 
 const DragedPlaylist = styled.div`
